// src/lib/types/Company.type.ts
export type Company = {
  id: string;
  ticker: string;
  name: string;
  legal_name: string;
  stock_exchange: string;
  short_description: string;
  long_description: string;
  company_url: string;
  business_address: string;
  business_phone_no: string;
  entity_legal_form: string;
  latest_filing_date: string;
  hq_country: string;
  employees: number;
  sector: string;
  industry_category: string;
  industry_group: string;
  thea_enabled: boolean;
  legacy_sector: string;
  legacy_industry_category: string;
  legacy_industry_group: string;
  first_stock_price_date: string;
  last_stock_price_date: string;
};